MacOS Edits
These are small edits made to the code to allow for Mac users to load the songs properly and to display album art properly. This should hopefully work with Windows users as well, but I am leaving them in this separate folder until further testing is done to confirm this.